[project @ 7be5f850f86a615643f44ab602cd411b3546cff5] Added more combat messages and disabled aggro and kill all for the moment
Original author: w4dimenscor <w4dimenscor>
Date: 2004-11-20 04:43:17+00:00

git-svn-id: svn+ssh://4dimensions.org/home/game-source/repositories/gamesrc@6 254f5b4c-bdf4-4503-907e-934c7b59e499

+++ act.informative.c
@@ -1343,8 +1343,9 @@ void look_at_room(struct char_data *ch, int ignore_brief)
   list_char_to_char(view_room->people, ch);
   new_send_to_char(ch, "%s", CCNRM(ch, C_NRM));
 
-  if ( PRF_FLAGGED(ch, PRF_AGGRO) )
+  if ( KILL_ALL_ENABLED && PRF_FLAGGED(ch, PRF_AGGRO) ) {
     command_interpreter(ch, "kill all");
+    }
 }
